pretrained/examples/ESPnet/inference/speech_to_text/requirements.txt
kenlm
numpy
espnet
scipy
espnet_model_zoo
soundfile
